create data module and move SimpleManipulating into there
