+++ .idea/gradle.xml
@@ -11,6 +11,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/data" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
