+++ app/src/main/java/black/bracken/picsorter/db/entity/SimpleManipulatingEntity.kt
@@ -3,7 +3,7 @@ package black.bracken.picsorter.db.entity
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import black.bracken.picsorter.service.model.SimpleManipulating
+import black.bracken.picsorter.data.SimpleManipulating
 
 @Entity(tableName = "manipulatings")
 data class SimpleManipulatingEntity(
@@ -13,7 +13,11 @@ data class SimpleManipulatingEntity(
 ) {
 
     fun toModel(): SimpleManipulating =
-        SimpleManipulating(name, newDirectoryPath, secondsToDelete)
+        SimpleManipulating(
+            name,
+            newDirectoryPath,
+            secondsToDelete
+        )
 
 }
