+++ settings.gradle.kts
@@ -1,1 +1,1 @@
-include(":app")
+include(":app", ":data")
